+++ Assets/Scripts/Managers/WinManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9276f7557e58b462b9db9e6513ca0b5c
