+++ Assets/Scripts/Puzzle/SnappingMethods.cs
@@ -1,31 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using Unity.VisualScripting;
 using UnityEngine;
 
-public class SnappingManager : MonoBehaviour
+public class SnappingMethods : MonoBehaviour
 {
-    [Header("Snapping Settings")]
-    public float snappingTolerance = 0.5f;
-
-    [Header("Animation Settings")]
-    public float snapSpeed = 0.25f; // the speed value of snapping
-    public AnimationCurve snapCurve = AnimationCurve.EaseInOut(0, 0, 1, 1); // the curve inside the inspector that controls the snapping curve
-
-    public ConnectionSystem connectionSystem;
-    private SnapValidation snapValidator = new SnapValidation();
-
-    [Header("Game")]
-    [SerializeField] private GameStateManager gameStateManager;
-
-    public bool IsAnimating { get; private set; } = false;
-
-    public struct SnapPairs
-    {
-        public PuzzlePiece draggedPiece;
-        public PuzzlePiece targetPiece;
-    }
     public void TrySnap(Transform pieceGroup)
     {
         if (connectionSystem == null)
@@ -141,135 +117,5 @@ public class SnappingManager : MonoBehaviour
             }
         }
         return false;
-    }
-
-    private List<SnapPairs> ScanConnections(Transform pieceGroup, Transform targetRoot, PuzzlePiece groupPiece, PuzzlePiece piece)
-    {
-        var scanConnections = new List<SnapPairs>();
-
-        foreach (var a in pieceGroup.GetComponentsInChildren<PuzzlePiece>())
-        {
-            foreach (var b in targetRoot.GetComponentsInChildren<PuzzlePiece>())
-            {
-                var distance = Vector2.Distance(a.SolvedPosition, b.SolvedPosition);
-                
-                if (distance > 0.1f && distance <= Vector2.Distance(groupPiece.SolvedPosition, piece.SolvedPosition) * 1.2f)
-                {
-                    var direction = (Vector2)a.SolvedPosition - (Vector2)b.SolvedPosition;
-                    if (Mathf.Abs(direction.x) < 0.1f || Mathf.Abs(direction.y) < 0.1f)
-                    {
-                        var connection = new SnapPairs { draggedPiece = a, targetPiece = b };
-                        scanConnections.Add(connection);
-                    }
-                }
-            }
-        }
-        
-        return scanConnections;
-    }
-    private IEnumerator Animate(Transform pieceGroup, Transform targetRoot, Vector3 start, Vector3 end, Quaternion startRotation, Quaternion endRotation, List<SnapPairs> scanConnections)
-    {
-        IsAnimating = true;
-
-        float elapsedTime = 0f;
-
-        Collider2D[] colliders = pieceGroup.GetComponentsInChildren<Collider2D>();
-
-        foreach (var collider in colliders)
-        {
-            collider.enabled = false;
-        }
-
-        while (elapsedTime < snapSpeed)
-        {
-            elapsedTime += Time.deltaTime;
-            pieceGroup.position = Vector3.Lerp(start, end, snapCurve.Evaluate(elapsedTime / snapSpeed));
-            pieceGroup.rotation = Quaternion.Lerp(startRotation, endRotation, snapCurve.Evaluate(elapsedTime / snapSpeed));
-            yield return null;
-        }
-
-        pieceGroup.position = end;
-        pieceGroup.rotation = endRotation;
-
-        MergeGroups(pieceGroup, targetRoot);
-
-        PlaySnapSound();
-
-        foreach (var connection in scanConnections)
-        {
-            Particles(connection.draggedPiece, connection.targetPiece);
-        }
-
-        RestorePieceRenderers(targetRoot);
-
-        foreach (var collider in colliders)
-        {
-            collider.enabled = true;
-        }
-
-        if (connectionSystem.Groups.IsPuzzleComplete())
-        {
-            gameStateManager.WinGame();
-        }
-
-        IsAnimating = false;
-    }
-    private Transform GetRoot(Transform piece)
-    {
-        while (piece.parent != null && piece.parent.CompareTag("Piece"))
-        {
-            piece = piece.parent;
-        }
-
-        return piece;
-    }
-
-    private bool IsValidSnapRotation(Transform pieceGroup, Transform targetPiece)
-    {
-        float angleA = Mathf.Round(pieceGroup.eulerAngles.z);
-        float angleB = Mathf.Round(targetPiece.eulerAngles.z);
-
-        float difference = Mathf.Abs(Mathf.DeltaAngle(angleA, angleB));
-
-        return difference <= 2f;
-    }
-
-    private void MergeGroups(Transform sourceRoot, Transform targetRoot)
-    {
-        if (sourceRoot == targetRoot)
-        {
-            return;
-        }
-
-        List<Transform> children = new List<Transform>();
-
-        foreach (Transform child in sourceRoot)
-        {
-            children.Add(child);
-        }
-
-        foreach (Transform child in children)
-        {
-            child.SetParent(targetRoot, true);
-        }
-
-        sourceRoot.SetParent(targetRoot, true);
-
-        if (targetRoot.GetComponent<PuzzlePiece>() != null)
-        {
-            foreach (var piece in targetRoot.GetComponentsInChildren<PuzzlePiece>())
-            {
-                if (piece.transform == targetRoot)
-                {
-                    continue;
-                }
-
-                Vector3 position = (Vector2)piece.SolvedPosition - (Vector2)targetRoot.GetComponent<PuzzlePiece>().SolvedPosition;
-                position.z = piece.transform.localPosition.z;
-
-                piece.transform.localPosition = position;
-                piece.transform.localRotation = Quaternion.identity;
-            }
-        }
     }
 }
