+++ Assets/Scripts/Rendering/AnimationController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 40f1852233edc4a08b9a1ec99668baf9
